# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SignalValidationOut": {
        "properties": {
          "classification_tags": {
            "items": {
              "type": "string"
            },
            "title": "Classification Tags",
            "type": "array"
          },
          "data_confidence_at_selection": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Confidence At Selection"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inclusion_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Inclusion Reasons",
            "type": "array"
          },
          "issue_codes": {
            "items": {
              "type": "string"
            },
            "title": "Issue Codes",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "provider_keys": {
            "items": {
              "type": "string"
            },
            "title": "Provider Keys",
            "type": "array"
          },
          "review_status": {
            "title": "Review Status",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          },
          "score_at_selection": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score At Selection"
          },
          "score_stratum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Stratum"
          },
          "signal_id": {
            "format": "uuid",
            "title": "Signal Id",
            "type": "string"
          },
          "signal_strength_at_selection": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Strength At Selection"
          },
          "source_keys": {
            "items": {
              "type": "string"
            },
            "title": "Source Keys",
            "type": "array"
          },
          "strength_stratum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strength Stratum"
          },
          "timing_at_selection": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timing At Selection"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          }
        },
        "required": [
          "id",
          "signal_id",
          "score_at_selection",
          "score_stratum",
          "signal_strength_at_selection",
          "data_confidence_at_selection",
          "timing_at_selection",
          "strength_stratum",
          "inclusion_reasons",
          "source_keys",
          "provider_keys",
          "classification_tags",
          "review_status",
          "verdict",
          "issue_codes",
          "notes",
          "reviewer",
          "reviewed_at"
        ],
        "title": "SignalValidationOut",
        "type": "object"
      },
...
      "ValidationBatchOut": {
        "properties": {
          "city_id": {
            "title": "City Id",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "definition_set_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Set Hash"
          },
          "dimension_scope_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimension Scope Key"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issue_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Issue Counts",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SignalValidationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pending_count": {
            "title": "Pending Count",
            "type": "integer"
          },
          "population_size": {
            "title": "Population Size",
            "type": "integer"
          },
          "reviewed_count": {
            "title": "Reviewed Count",
            "type": "integer"
          },
          "rule_version": {
            "title": "Rule Version",
            "type": "string"
          },
          "score_as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score As Of Date"
          },
          "score_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Run Id"
          },
          "seed": {
            "title": "Seed",
            "type": "string"
          },
          "selection_manifest": {
            "additionalProperties": true,
            "title": "Selection Manifest",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_size": {
            "title": "Target Size",
            "type": "integer"
          },
          "verdict_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Verdict Counts",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "city_id",
          "rule_version",
          "dimension_scope_key",
          "score_run_id",
          "score_as_of_date",
          "definition_set_hash",
          "seed",
          "target_size",
          "population_size",
          "status",
          "reviewed_count",
          "pending_count",
          "verdict_counts",
          "issue_counts",
          "created_at",
          "completed_at",
          "selection_manifest",
          "items"
        ],
        "title": "ValidationBatchOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FlyttSignal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/validation-batches/{batch_id}": {
      "get": {
        "operationId": "getValidationBatch",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationBatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Validation Batch"
      }
    }
  }
}